# zema_store_backend/src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zema Store API with Swagger",
        "version": "1.0.0",
        "description": "Zema Store API",
        "contact": {
            "name": "Zema Store",
            "url": "http://zema-store.herokuapp.com",
            "email": "henokadane8@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://zema-store.herokuapp.com/api"
        },
        {
            "url": "http://localhost:3000/api"
        }
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Returns list of songs",
                "description": "Returns list of all songs",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "type": "integer",
                        "description": "The page number of song items to return"
                    },
                    {
                        "in": "query",
                        "name": "sortBy",
                        "type": "string",
                        "description": "song object field to sort returned songs",
                        "enum": [
                            "title",
                            "releaseDate",
                            "createdAt"
                        ],
                        "example": "title:asc"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Adds a song",
                "description": "Adds a song",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SongRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Returns song data",
                "description": "Returns song data with specified id",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified song"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Songs"
                ],
                "summary": "Updates song data",
                "description": "Updates song data with specified id",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified song"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SongRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Deletes song data",
                "description": "Deletes song data specified by id",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified song"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuthorization": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token should be accessToken for user"
            }
        },
...
        "schemas": {
            "SuccessMessage": {
                "type": "string",
                "description": "Success message",
                "nullable": true,
                "required": true
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message",
                        "required": true
                    },
                    "data": {
                        "type": "object",
                        "required": true,
                        "nullable": true
                    }
                }
            },
            "OkResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "$ref": "#/components/schemas/SuccessMessage"
                    },
                    "data": {
                        "type": "object",
                        "required": true,
                        "nullable": false
                    }
                },
                "required": [
                    "message",
                    "data"
                ]
            },
...
            "SongRequest": {
                "type": "object",
                "properties": {
                    "albumId": {
                        "type": "string",
                        "description": "Album Id"
                    },
                    "artistId": {
                        "type": "string",
                        "description": "Artist Id"
                    },
                    "title": {
                        "type": "string",
                        "description": "Song title"
                    },
                    "song": {
                        "type": "string",
                        "format": "binary",
                        "description": "Song file"
                    },
                    "genre": {
                        "type": "string",
                        "description": "Song genre",
                        "enum": [
                            "reggae",
                            "pop",
                            "rock"
                        ]
                    },
                    "length": {
                        "type": "number",
                        "description": "Song duration"
                    },
                    "releaseDate": {
                        "type": "string",
                        "description": "Song release date",
                        "format": "date-time"
                    }
                },
                "required": [
                    "albumId",
                    "artistId",
                    "title",
                    "song",
                    "length",
                    "releaseDate"
                ]
            }
        }
    }
}